Delete `bugzilla` module, move all models to `models.py`
