+++ tests/unit/test_bugzilla.py
@@ -4,8 +4,8 @@ Module for testing jbi/configuration.py functionality
 # pylint: disable=cannot-enumerate-pytest-fixtures
 import pytest
 
-from jbi import bugzilla
 from jbi.errors import ActionNotFoundError
+from jbi.models import BugzillaBug
 
 
 @pytest.mark.parametrize(
@@ -29,7 +29,7 @@ from jbi.errors import ActionNotFoundError
     ],
 )
 def test_jira_labels(whiteboard, expected):
-    bug = bugzilla.BugzillaBug(id=0, whiteboard=whiteboard)
+    bug = BugzillaBug(id=0, whiteboard=whiteboard)
     assert bug.get_jira_labels() == expected
 
 
@@ -40,7 +40,7 @@ def test_jira_labels(whiteboard, expected):
         (["foo"], None),
         (["fail:/format"], None),
         (["foo", "http://jira.net/123"], "123"),
-        (["http://bugzilla.org/123"], None),
+        (["http://org/123"], None),
         (["http://jira.com"], None),
         (["http://mozilla.jira.com/"], None),
         (["http://mozilla.jira.com/123"], "123"),
@@ -51,21 +51,19 @@ def test_jira_labels(whiteboard, expected):
     ],
 )
 def test_extract_see_also(see_also, expected):
-    test_bug = bugzilla.BugzillaBug(id=0, see_also=see_also)
+    test_bug = BugzillaBug(id=0, see_also=see_also)
     assert test_bug.extract_from_see_also() == expected
 
 
 def test_lookup_action(actions_example):
-    bug = bugzilla.BugzillaBug.parse_obj(
-        {"id": 1234, "whiteboard": "[example][DevTest]"}
-    )
+    bug = BugzillaBug.parse_obj({"id": 1234, "whiteboard": "[example][DevTest]"})
     action = bug.lookup_action(actions_example)
     assert action.whiteboard_tag == "devtest"
     assert "test config" in action.description
 
 
 def test_lookup_action_missing(actions_example):
-    bug = bugzilla.BugzillaBug.parse_obj({"id": 1234, "whiteboard": "example DevTest"})
+    bug = BugzillaBug.parse_obj({"id": 1234, "whiteboard": "example DevTest"})
     with pytest.raises(ActionNotFoundError) as exc_info:
         bug.lookup_action(actions_example)
     assert str(exc_info.value) == "example devtest"
