+++ tests/unit/test_main.py
@@ -5,7 +5,7 @@ import pytest
 from fastapi.testclient import TestClient
 
 from jbi.app import app
-from jbi.bugzilla import BugzillaWebhookRequest
+from jbi.models import BugzillaWebhookRequest
 
 
 def test_request_summary_is_logged(caplog):
